examples/requirements.txt
lightrag-hku>=1.3.0
lightrag-spanner @ git+https://github.com/ksmin23/lightrag-spanner.gitex